Changing lexer from Sequences to a custom Generator implementation. I was trying to do this similar to how it was done in Uniform (https://github.com/uniform-team/Uniform-Validation-Language/blob/master/src/lexer.js), but Sequences do not work the way I expected.
It appears that Sequence<T>#drop(int), forces a read through the sequence in a chained manner. As a result, continuously dropping 1 character at a time, meant that it was reading the file all the way through, each time. This is needlessly excessive and I couldn't find a good way around it, so I built out a JavaScript-like Generator and did it that way.

There's probably a way to do this will Kotlin's yield functionality, but this was the simpler option for me.