Adding initial file reading and lexer setup.
If I were a more competent C++ developer, this would not load everything into std::queue, but instead use some kind of coroutine to parallelize the file reading, lexical analysis, and parsing to run simultaneously without loading everything into memory at once.

Also updating the project configuration to enforce the style that pointer declarations are associated with the type (which I prefer because I'm a heathen and don't declare multiple variables in the same statement). Also adding a clang-tidy exception for modernize-pass-by-value because switching it to a std::move(...) causes more problems than it solves.