Adding GoogleTest and tests for existing code.
The test system is mostly working. Some extra files are built in the main source directory which I can't get into the build/ directory which is annoying but not the end of the world. If I were better at CMake I could probably get that working. Intellij also has a special GoogleTest runner which I can't get working at the moment. Running the binary is good enough for now, but I'll want to come back to that later.

clang-tidy also has some warnings on the TEST() lines, but adding the modernize-use-equals-delete exception doesn't seem to work, and I can't find the name of the other warning.