Adding basic parsing.
This adds a very simpler parser which is capable of parsing statements consisting only of functional calls to functions which accept a single character as their parameter. This should be enough for putchar() to perform a basic "Hello World!", more complex parsing can come after that.

I've also hit the point of the C++ development lifecycle where I get various linker errors which make no sense. As far as I can tell, they seem to be some kind of dependency/ordering issue. I refactored the CMakeLists.txt file to make each file its own library with its own listed dependencies. I like this a lot better for dependency management, but it seems that a lot of dependencies can be included transitively, so it may not be all that useful. Regardless, it seemed to fix the particular errors I was running into, so I'm going to stick with that.

I also refactored the test system, giving each of them their own targets and linking them individually with their dependencies. However I couldn't get the test runner to actually execute all of them in an easy fashion. They need to be added as dependencies to the test runner and then #include'd in the test runner's C++ code, which definitely shouldn't be necessary. I'll need to spend more time to figure out a more scalable test system.