Adding line number and column tracking to lexer.
Each Token is now generated with line and a column span which can be used for more accurate and pointed error messages. Tokens only have a single line number, but a column start and end because in Sanity, Tokens cannot cross a newline, so having a Token on two lines is impossible.