Hello World!
Got a "Hello World!" Sanity program to run! Added a Generator class to visited all the AST nodes and generate the appropriate LLVM constructs from them to create the IR source code and print it out. This can be most easily run with:

$ cmake --build . --target Sanity && build/bin/Sanity hello.sane | lli

`lli` is the LLVM interpreter for the IR. It is also possible to build an actual binary.

$ cmake --build . --target Sanity && build/bin/Sanity hello.sane | llc | g++ -x assembler - -o  build/bin/hello && build/bin/hello

`llc` is the LLVM compiler which transpiles IR into assembly for the current architecture.
`g++` is used to finally assemble and link the code with the standard C library (where putchar() is defined).

We really should be having `llc` output object code and then use `ld` to link it with the standard C library. By using `g++` we're mixing toolchains somewhat. Unfortunately, I couldn't get the `ld` to link correctly, it appears to be accessing a shared library which doesn't exist on my machine, but `g++` links correctly without any extra effort, so I'm going with that for the moment.

Also, gotta love the hard newline character in hello.sane. The lexer isn't capable of handling escape characters, but also doesn't have anything blacklisted, so any character literals are accepted, including newline.

For simplicity, I basically assumed that the only actions which can be performed are to use an `extern` to link against a function defined externally as well invoke a function which accepts a 32-bit integer and returns one. For some reason `putchar()` accepts a 32-bit integer even though it's meant for chars which are 8-bits by default.

The next few commits will clean things up and make things a little more flexible and less hacky. Then I'll start adding to the lexer / parser / generator to slowly add features one at a time.