Updating `extern` statement with proper type syntax.
Starting to remove some of the assumptions I used to simplify the "Hello World!" program. Here I'm expanding functions to accept more than just a single character and returning more than a single character. Of course, the only option is `int`, so this doesn't change a whole lot, but the parsing is more flexible and accurate to reality.

In testing, there is some state bleeding through, notably a function can be defined in one test and then called in another which is a problem. The module needs to be reset but I can't find a method of doing that. For now I've just avoided name conflicts, but this is definitely something that needs to be fixed sooner rather than later.