Simplifying error message creation.
After speaking with someone who actually knows C++, it seems that the std::string#c_str() function returns a direct reference to its inner buffer, rather than allocating new memory. That means that when the std::string object is deleted, so is the char* it previously returned.

Best way around this particular problem is to store the string as an instance variable of the exception class, that means any c_str() call's result will stick around as long as that exception is in memory, which is at least for the duration of the catch block. I updated the exceptions to create the string in the constructor rather than the what() function.

Thanks for the help Rich!