Adding GFlags library.
Added support for a --input flag to set the input filename. If not specified, the compiler will read from stdin instead.

Note: This will break on Windows because it hard codes /dev/stdin, which won't work for Windows. Since Windows was only sort-of supported previously, I don't use it for development, there's no one else working on this, and because I just don't care, I simply documented this "bug" as a feature. Problem solved.