Switching build system from CMake to Bazel.
CMake had a few issues:
1) Difficult to integrate with a custom compiler.
2) External dependencies had to be manually downloaded.
3) Required a special run.sh command just to execute.
4) Tests could not easily depend on the compiler.
5) 100% of the one-person development team was very unfamiliar with it.

Bazel solves most of these issues:
1) I added a `sanity_binary()` rule to handle compilation of Sanity code in an abstracted form.
2) External dependencies are handled automatically by Bazel.
3) Bazel is powerful enough not to require a special run script.
4) Tests can depend on the compiler or a compiler output binary.
5) The entire development team is much more familiar with Bazel.

I did this all in one commit because I'm the only developer and it's still small enough that I can get away with this. Integrating Bazel with LLVM was the trickiest part, but that was largely solved in https://github.com/dgp1130/llvm-bazel-foolang and was copied for this project.

Some minor changes needed to be made with the Generator. I had to make an IGenerator interface in AST in order to handle some circular dependency issues. I also had to define the global LLVM data which were not necessary before for some reason.

I also removed all end-to-end tests for the moment. They manually compiled source code to check compiler and result output. Bazel handles all the building now, so the Python test should now. I intend to update these to build with Bazel and then simply assert the output.