Adding back end-to-end tests for the Sanity compiler.
Created build definitions and a test wrapper to execute a given program and assert its outputs. This gets a little repetitive with the arithmetic tests because there's no easy way to execute a single program with multiple test cases. Could put more effort into the `test_sanity_prog()` macro so it can handle proper test cases, but I didn't want to deal with that right now.