Adding comment lexical analysis.
Sanity now supports standard C-like comments such as:
// Single-line comments like this.
/* Multi-line comments
   like this */

In the course of debugging this I realized that the Hello run configuration didn't work with the debugger because the Sanity compiler is executed as a build step. I added a "Hello Debug" run configuration which just executes the compiler alone and should work with the debugger.