Fixing cat test to test what it intends to.
When this test was implemented it was having issues and I had to debug it by slowing converting into into the printf test. Eventually I solved the problem, but forgot to switch the test back to actually implementing the basic interface of the cat utility. Updated the test to reflect what it was actually intended to test.