Adding basic variable bindings.
Currently the syntax to use this the following:
let foo: int = 0;
foo; // 0

This works for integers and strings (chars are technically integers under the hood, so they're a little weird). It throws errors if the given name is declared multiple times, if the name is used before it is declared, or if the type of the binding does not match the value it is initialized with. A binding cannot be initialized without a starting value.

These bindings are also immutable and cannot be changed (hence why I'm avoiding the word "variable"). There is no current grammar which would enable a write operation over an existing binding.

This only has very basic type checking at the LLVM abstraction level which does not provide very good output. Next thing is to add a simple type checker at the Sanity AST level which can provide helpful error messages.

In doing that, it would also be necessary to propagate type information up the expression tree at the Sanity AST level which would enable type inference fairly trivially.